+++ dlt/destinations/bigquery/__init__.py
@@ -18,10 +18,10 @@ def _configure(config: BigQueryClientConfiguration = config.value) -> BigQueryCl
 def capabilities() -> DestinationCapabilitiesContext:
     caps = DestinationCapabilitiesContext()
     caps.preferred_loader_file_format = "jsonl"
-    caps.supported_loader_file_formats = ["jsonl", "sql", "parquet"]
-    caps.preferred_staging_file_format = None
-    caps.supported_staging_file_formats = []
-    caps.supported_stagings = []
+    caps.supported_loader_file_formats = ["jsonl", "sql", "parquet", "reference"]
+    caps.preferred_staging_file_format = "parquet"
+    caps.supported_staging_file_formats = ["parquet"]
+    caps.supported_stagings = ["filesystem"]
     caps.escape_identifier = escape_bigquery_identifier
     caps.escape_literal = None
     caps.max_identifier_length = 1024
