+++ dlt/destinations/insert_job_client.py
@@ -5,13 +5,10 @@ from typing import Any, Iterator, List, Type
 from dlt.common.destination.reference import LoadJob, FollowupJob, TLoadJobState
 from dlt.common.schema.typing import TTableSchema, TWriteDisposition
 from dlt.common.storages import FileStorage
-from dlt.common.configuration import with_config, known_sections
-from dlt.common.configuration.accessors import config
 
 from dlt.destinations.sql_client import SqlClientBase
 from dlt.destinations.job_impl import EmptyLoadJob
 from dlt.destinations.job_client_impl import SqlJobClientBase
-from dlt.destinations.filesystem.configuration import FilesystemClientConfiguration
 
 
 class InsertValuesLoadJob(LoadJob, FollowupJob):
@@ -71,26 +68,6 @@ class InsertValuesLoadJob(LoadJob, FollowupJob):
         if insert_sql:
             yield insert_sql
 
-@with_config(spec=FilesystemClientConfiguration, sections=(known_sections.DESTINATION, "filesystem",))
-def _fs_configure(config: FilesystemClientConfiguration = config.value) -> FilesystemClientConfiguration:
-    return config
-
-class CopyFileLoadJob(LoadJob, FollowupJob):
-    def __init__(self, table_name: str, file_path: str, sql_client: SqlClientBase[Any]) -> None:
-        super().__init__(FileStorage.get_file_name_from_file_path(file_path))
-        self._sql_client = sql_client
-        with open(file_path, "r+", encoding="utf-8") as f:
-            # Reading from a file
-            bucket_path = f.read()
-        self.execute(table_name, bucket_path, _fs_configure(FilesystemClientConfiguration(dataset_name="something")))
-
-    def execute(self, table_name: str, bucket_path: str, fs_config: FilesystemClientConfiguration) -> None:
-        # implement in child implementations
-        return
-
-    def state(self) -> TLoadJobState:
-        # this job is always done
-        return "completed"
 
 class InsertValuesJobClient(SqlJobClientBase):
 
@@ -113,9 +90,6 @@ class InsertValuesJobClient(SqlJobClientBase):
 
     def start_file_load(self, table: TTableSchema, file_path: str, load_id: str) -> LoadJob:
         job = super().start_file_load(table, file_path, load_id)
-        if file_path.endswith(".reference"):
-            JobClass = self.get_file_copy_job()
-            job = JobClass(table["name"], file_path, self.sql_client)
         if not job:
             # this is using sql_client internally and will raise a right exception
             job = InsertValuesLoadJob(table["name"], table["write_disposition"], file_path, self.sql_client)
@@ -130,6 +104,3 @@ class InsertValuesJobClient(SqlJobClientBase):
         # set non unique indexes
         pass
 
-    @staticmethod
-    def get_file_copy_job() -> Type[CopyFileLoadJob]:
-        return CopyFileLoadJob
