+++ dlt/destinations/redshift/redshift.py
@@ -15,8 +15,9 @@ from dlt.common.destination import DestinationCapabilitiesContext
 from dlt.common.data_types import TDataType
 from dlt.common.schema import TColumnSchema, TColumnHint, Schema
 
-from dlt.destinations.insert_job_client import InsertValuesJobClient, CopyFileLoadJob
+from dlt.destinations.insert_job_client import InsertValuesJobClient
 from dlt.destinations.exceptions import DatabaseTerminalException
+from dlt.destinations.job_client_impl import CopyFileLoadJob
 
 from dlt.destinations.redshift import capabilities
 from dlt.destinations.redshift.configuration import RedshiftClientConfiguration
